requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
pydantic==2.4.2
openai==1.3.0
langchain==0.0.330
langchain-openai==0.0.2
pytest==7.4.3
pytest-flask==1.3.0
black==23.9.1
flake8==6.1.0

# Document parsing with Docling
docling>=2.34.0

# Supporting libraries
pandas==2.1.3
pillow==10.0.1
opencv-python==4.8.1.78
numpy==1.24.3
requests==2.31.0

# Additional document processing
python-docx==0.8.11
openpyxl==3.1.2
xlrd==2.0.1